+++ src/test/java/com/koddy/server/acceptance/auth/TokenReissueAcceptanceTest.java
@@ -1,7 +1,7 @@
 package com.koddy.server.acceptance.auth;
 
 import com.koddy.server.common.AcceptanceTest;
-import com.koddy.server.common.config.DatabaseCleanerEachCallbackExtension;
+import com.koddy.server.common.containers.callback.DatabaseCleanerEachCallbackExtension;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
