+++ src/test/java/com/koddy/server/common/AcceptanceTest.java
@@ -8,8 +8,8 @@ import com.koddy.server.auth.infrastructure.oauth.zoom.ZoomOAuthConnector;
 import com.koddy.server.auth.infrastructure.oauth.zoom.ZoomOAuthUriGenerator;
 import com.koddy.server.coffeechat.infrastructure.link.zoom.ZoomMeetingLinkManager;
 import com.koddy.server.common.config.ExternalApiConfiguration;
-import com.koddy.server.common.config.MySqlTestContainersExtension;
-import com.koddy.server.common.config.RedisTestContainersExtension;
+import com.koddy.server.common.containers.MySqlTestContainersExtension;
+import com.koddy.server.common.containers.RedisTestContainersExtension;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
