+++ src/test/java/com/koddy/server/common/IntegrateTest.java
@@ -1,8 +1,8 @@
 package com.koddy.server.common;
 
-import com.koddy.server.common.config.DatabaseCleanerEachCallbackExtension;
-import com.koddy.server.common.config.MySqlTestContainersExtension;
-import com.koddy.server.common.config.RedisTestContainersExtension;
+import com.koddy.server.common.containers.MySqlTestContainersExtension;
+import com.koddy.server.common.containers.RedisTestContainersExtension;
+import com.koddy.server.common.containers.callback.DatabaseCleanerEachCallbackExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
