+++ src/test/java/com/koddy/server/common/RedisTest.java
@@ -1,6 +1,6 @@
 package com.koddy.server.common;
 
-import com.koddy.server.common.config.RedisTestContainersExtension;
+import com.koddy.server.common.containers.RedisTestContainersExtension;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.autoconfigure.data.redis.DataRedisTest;
