+++ src/test/java/com/koddy/server/common/containers/callback/RedisCleanerAllCallbackExtension.java
@@ -1,4 +1,4 @@
-package com.koddy.server.common.config;
+package com.koddy.server.common.containers.callback;
 
 import com.koddy.server.common.utils.RedisCleaner;
 import org.junit.jupiter.api.extension.AfterAllCallback;
